server/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# CORS and middleware
starlette==0.27.0

# Grammar checking
language-tool-python==2.7.1

# Text rewriting and AI models
transformers==4.36.0
torch==2.1.0
accelerate==0.25.0
sentencepiece==0.1.99

# Audio processing and transcription
faster-whisper==0.10.0
ffmpeg-python==0.2.0

# Tone analysis (optional ONNX support)
onnxruntime==1.16.0
scikit-learn==1.3.2
nltk==3.8.1

# Utilities
requests==2.31.0
numpy==1.24.4
regex==2023.10.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
